feat: 调整search的整体接口
