+++ binding/rust/ip2region2/Cargo.toml
@@ -10,6 +10,8 @@ license = "Apache-2.0"
 
 [dependencies]
 once_cell = "1.16"
+tracing = "0.1"
+#tracing-subscriber = "0.2"
 
 [dev-dependencies]
 criterion = "0.4"
